Renames `diags` package to `diag`
